Fix old read_thread may never exit when gamepad updates in windows
In update_gamepads() in the case of an updated gamepad, alive is
set to false to signal to the old thread to exit out of its while
loop.

But there was no join() so, if the old thread is too slow, when
gamepads[joy_id] is updated, the old thread will receive alive=true
and stay in it's loop.